refactor(package): restructure base project
