+++ config/webpack.common.js
@@ -16,7 +16,7 @@ const AssetsPlugin = require('assets-webpack-plugin');
  */
 const HMR = helpers.hasProcessFlag('hot');
 const METADATA = {
-  title: 'Angular2',
+  title: 'Terram Game',
   baseUrl: '/',
   isDevServer: helpers.isWebpackDevServer()
 };
