--- src/app/+detail/detail.component.ts
@@ -1,19 +1,0 @@
-import { Component } from '@angular/core';
-
-@Component({
-  selector: 'detail',
-  template: `
-    <h1>Hello from Detail</h1>
-    <router-outlet></router-outlet>
-  `
-})
-export class Detail {
-  constructor() {
-
-  }
-
-  ngOnInit() {
-    console.log('hello `Detail` component');
-  }
-
-}
