--- src/app/about/index.ts
@@ -1,1 +1,0 @@
-export * from './about.component';
