--- src/app/app.style.css
@@ -1,8 +1,0 @@
-html, body{
-  height: 100%;
-  font-family: Arial, Helvetica, sans-serif
-}
-
-span.active {
-  background-color: gray;
-}
