+++ src/app/components/terram/terram.component.ts
@@ -1,0 +1,10 @@
+import { Component, ViewEncapsulation } from '@angular/core';
+
+@Component({
+  selector: 'terram',
+  encapsulation: ViewEncapsulation.None,
+  template: require('./terram.html'),
+})
+export class Terram {
+  constructor() {}
+}
