+++ src/app/components/terram/terram.e2e.ts
@@ -1,4 +1,4 @@
-describe('App', () => {
+describe('terram', () => {
 
   beforeEach(() => {
     browser.get('/');
@@ -18,15 +18,8 @@ describe('App', () => {
   });
 
   it('should have <home>', () => {
-    let subject = element(by.css('app home')).isPresent();
+    let subject = element(by.css('terram home')).isPresent();
     let result  = true;
     expect(subject).toEqual(result);
   });
-
-  it('should have buttons', () => {
-    let subject = element(by.css('button')).getText();
-    let result  = 'Submit Value';
-    expect(subject).toEqual(result);
-  });
-
 });
