+++ src/app/components/terram/terram.html
@@ -1,0 +1,29 @@
+<nav>
+  <span>
+    <a [routerLink]=" ['./'] ">
+      Index
+    </a>
+  </span>
+  |
+  <span>
+    <a [routerLink]=" ['./home'] ">
+      Home
+    </a>
+  </span>
+  |
+  <span>
+    <a [routerLink]=" ['./detail'] ">
+      Detail
+    </a>
+  </span>
+  |
+  <span>
+    <a [routerLink]=" ['./about'] ">
+      About
+    </a>
+  </span>
+</nav>
+
+<main>
+<router-outlet></router-outlet>
+</main>
