--- src/app/home/title/index.ts
@@ -1,1 +1,0 @@
-export * from './title.service';
