--- src/app/home/title/title.service.ts
@@ -1,19 +1,0 @@
-import { Injectable } from '@angular/core';
-import { Http } from '@angular/http';
-
-@Injectable()
-export class Title {
-  value = 'Angular 2';
-  constructor(public http: Http) {
-
-  }
-
-  getData() {
-    console.log('Title#getData(): Get Data');
-    // return this.http.get('/assets/data.json')
-    // .map(res => res.json());
-    return {
-    };
-  }
-
-}
