+++ src/app/index.ts
@@ -1,2 +1,2 @@
 // App
-export * from './app.module';
+export * from './components/terram/terram.module';
