+++ src/app/pages/home/home.spec.ts
@@ -2,7 +2,6 @@ import {
   inject,
   TestBed
 } from '@angular/core/testing';
-import { Component } from '@angular/core';
 import {
   BaseRequestOptions,
   ConnectionBackend,
@@ -11,9 +10,7 @@ import {
 import { MockBackend } from '@angular/http/testing';
 
 // Load the implementations that should be tested
-import { AppState } from '../app.service';
 import { Home } from './home.component';
-import { Title } from './title';
 
 describe('Home', () => {
   // provide our implementations or mocks to the dependency injector
@@ -28,18 +25,8 @@ describe('Home', () => {
         },
         deps: [MockBackend, BaseRequestOptions]
       },
-      AppState,
-      Title,
       Home
     ]
-  }));
-
-  it('should have default data', inject([ Home ], (home: Home) => {
-    expect(home.localState).toEqual({ value: '' });
-  }));
-
-  it('should have a title', inject([ Home ], (home: Home) => {
-    expect(!!home.title).toEqual(true);
   }));
 
   it('should log ngOnInit', inject([ Home ], (home: Home) => {
