+++ src/app/pages/home/home.template.html
@@ -1,0 +1,10 @@
+<div class="card-container">
+  <h1 x-large class="sample-content">Your Content Here</h1>
+
+  <hr>
+
+  <div>
+    For hot module reloading run
+    <pre>npm run start:hmr</pre>
+  </div>
+</div>
