+++ src/app/pages/home/x-large/x-large.directive.ts
@@ -1,4 +1,4 @@
-import { Component, Directive, ElementRef, Renderer } from '@angular/core';
+import { Directive, ElementRef, Renderer } from '@angular/core';
 /*
  * Directive
  * XLarge is a simple directive to show how one is made
