+++ src/app/pages/home/x-large/x-large.spec.ts
@@ -1,13 +1,10 @@
 import {
   fakeAsync,
-  inject,
   tick,
   TestBed
 } from '@angular/core/testing';
 import { Component } from '@angular/core';
-import { BaseRequestOptions, Http } from '@angular/http';
 import { By } from '@angular/platform-browser/src/dom/debug/by';
-import { MockBackend } from '@angular/http/testing';
 
 // Load the implementations that should be tested
 import { XLarge } from './x-large.directive';
