+++ src/main.browser.ts
@@ -2,20 +2,21 @@
  * Angular bootstraping
  */
 import { platformBrowserDynamic } from '@angular/platform-browser-dynamic';
-import { decorateModuleRef } from './app/environment';
+import { decorateModuleRef } from './app/components/terram/environment';
 import { bootloader } from '@angularclass/hmr';
+
 /*
  * App Module
  * our top level module that holds all of our components
  */
-import { AppModule } from './app';
+import { TerramModule } from './app';
 
 /*
  * Bootstrap our Angular app with a top level NgModule
  */
 export function main(): Promise<any> {
   return platformBrowserDynamic()
-    .bootstrapModule(AppModule)
+    .bootstrapModule(TerramModule)
     .then(decorateModuleRef)
     .catch(err => console.error(err));
 }
